W-15426401: generalize reference handling in async parsers
