+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/async/parser/bindings/BindingParser.scala
@@ -3,6 +3,7 @@ package amf.apicontract.internal.spec.async.parser.bindings
 import amf.apicontract.client.scala.model.domain.bindings.BindingVersion
 import amf.apicontract.internal.spec.async.parser.context.AsyncWebApiContext
 import amf.apicontract.internal.spec.common.parser.SpecParserOps
+import amf.apicontract.internal.spec.spec.OasDefinitions
 import amf.apicontract.internal.validation.definitions.ParserSideValidations.{
   UnsupportedBindingVersion,
   UnsupportedBindingVersionWarning
@@ -10,14 +11,15 @@ import amf.apicontract.internal.validation.definitions.ParserSideValidations.{
 import amf.core.client.scala.model.domain.{AmfElement, AmfObject, AmfScalar, DomainElement}
 import amf.core.internal.metamodel.Field
 import amf.core.internal.parser.YMapOps
-import amf.core.internal.parser.domain.Annotations
+import amf.core.internal.parser.domain.{Annotations, SearchScope}
 import amf.core.internal.remote.Spec
 import amf.core.internal.remote.Spec._
+import amf.core.internal.validation.CoreValidations
 import amf.shapes.internal.spec.common.JSONSchemaDraft7SchemaVersion
 import amf.shapes.internal.spec.common.parser.YMapEntryLike
 import amf.shapes.internal.spec.oas.parser.OasTypeParser
 import amf.shapes.internal.validation.definitions.ShapeParserSideValidations.RequiredField
-import org.yaml.model.{YMap, YMapEntry, YNode, YNodePlain, YPart, YScalar, YSequence}
+import org.yaml.model._
 
 trait BindingParser[+Binding <: DomainElement] extends SpecParserOps {
 
@@ -58,13 +60,13 @@ trait BindingParser[+Binding <: DomainElement] extends SpecParserOps {
 
   protected def getDefaultBindingVersion(binding: String, spec: Spec): String = {
     (binding, spec) match {
-      case ("Amqp091ChannelBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26)   => "0.1.0"
-      case ("Amqp091OperationBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26) => "0.1.0"
-      case ("Amqp091MessageBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26)   => "0.1.0"
-      case("AnypointMQMessageBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26) => "0.0.1"
-      case("AnypointMQChannelBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26) => "0.0.1"
-      case ("KafkaOperationBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26)   => "0.1.0"
-      case ("KafkaMessageBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26)     => "0.1.0"
+      case ("Amqp091ChannelBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26)    => "0.1.0"
+      case ("Amqp091OperationBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26)  => "0.1.0"
+      case ("Amqp091MessageBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26)    => "0.1.0"
+      case ("AnypointMQMessageBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26) => "0.0.1"
+      case ("AnypointMQChannelBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26) => "0.0.1"
+      case ("KafkaOperationBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26)    => "0.1.0"
+      case ("KafkaMessageBinding", ASYNC20 | ASYNC21 | ASYNC22 | ASYNC23 | ASYNC24 | ASYNC25 | ASYNC26)      => "0.1.0"
     }
   }
 
@@ -106,7 +108,7 @@ trait BindingParser[+Binding <: DomainElement] extends SpecParserOps {
   ): Unit = {
     val entryLike = entry match {
       case map: YMapEntry => YMapEntryLike(map.value)
-      case node: YNode => YMapEntryLike(node)
+      case node: YNode    => YMapEntryLike(node)
     }
     OasTypeParser(
       entryLike,
@@ -119,6 +121,34 @@ trait BindingParser[+Binding <: DomainElement] extends SpecParserOps {
         binding.setWithoutId(field, shape, Annotations(entry))
         shape
       }
+  }
+
+  protected def handleRef(fullRef: String, searchLabel: String, entry: YMapEntry, field: Field, binding: DomainElement)(
+      implicit ctx: AsyncWebApiContext
+  ): Unit = {
+    val label = OasDefinitions.stripOas3ComponentsPrefix(fullRef, searchLabel)
+    ctx.declarations
+      .findType(label, SearchScope.Named)
+      .map(shape => binding.setWithoutId(field, shape, Annotations(entry)))
+      .getOrElse {
+        remote(fullRef, entry, field, binding)
+      }
+  }
+
+  private def remote(fullRef: String, entry: YMapEntry, field: Field, binding: DomainElement)(implicit
+      ctx: AsyncWebApiContext
+  ): Unit = {
+    ctx.navigateToRemoteYNode(fullRef) match {
+      case Some(remoteResult) =>
+        parseSchema(field, binding, remoteResult.remoteNode)
+      case None =>
+        ctx.eh.violation(
+          CoreValidations.UnresolvedReference,
+          binding,
+          s"Cannot find link reference $fullRef",
+          entry.location
+        )
+    }
   }
 
   protected def missingRequiredFieldViolation(
